Edit now works

+++ pages/models.py
@@ -42,14 +42,14 @@ class Project(models.Model):
         MinValueValidator(0), MaxValueValidator(100)])
     cash_discount = models.IntegerField(null=True, blank=True, default=0)
     discount_total = models.IntegerField(null=True, blank=True, default=0)
-    final_total = models.IntegerField(null=True, blank=True, default=0)
-    deposit_percentage = models.IntegerField(null=True, blank=True, default=0, validators=[
+    final_total = models.FloatField(null=True, blank=True, default=0)
+    deposit_percentage = models.FloatField(null=True, blank=True, default=0, validators=[
         MaxValueValidator(100), MinValueValidator(0)])
-    deposit_amount = models.IntegerField(null=True, blank=True, default=0)
-    completion_percentage = models.IntegerField(
+    deposit_amount = models.FloatField(null=True, blank=True, default=0)
+    completion_percentage = models.FloatField(
         null=True, blank=True, default=0, validators=[
             MaxValueValidator(100), MinValueValidator(0)])
-    completion_amount = models.IntegerField(null=True, blank=True, default=0, )
+    completion_amount = models.FloatField(null=True, blank=True, default=0, )
     intial_date = models.DateField(
         default=timezone.now, auto_now_add=False, auto_now=False, blank=True)
     survey_date = models.DateField(null=True,
